@@ -41,7 +41,7 @@ func SetToken(w http.ResponseWriter, r *http.Request) {
 
 // SendPayload is a wrapper for converting the payload map parameter into json and
 // sending to the client
-func SendPayload(w http.ResponseWriter, payload interface{}, errResp HTTPResponseConfig) error {
+func SendPayload(w http.ResponseWriter, payload any, errResp HTTPResponseConfig) error {
 	w.Header().Set("Content-Type", JSON_CONTENT_HEADER)
 	SetHTTPResponseDefaults(&errResp, http.StatusInternalServerError, []byte(invalidJSONTxt))
 	jsonString, err := json.Marshal(payload)
@@ -90,7 +90,7 @@ func DecodeCookieNoRequest(cookieName, val string, authKey, encryptKey []byte) (
 }
 
 // GetJSONBuffer takes interface and json encodes it into a buffer and returns buffer
-func GetJSONBuffer(item interface{}) (bytes.Buffer, error) {
+func GetJSONBuffer(item any) (bytes.Buffer, error) {
 	var buffer bytes.Buffer
 
 	encoder := json.NewEncoder(&buffer)
