@@ -5,8 +5,12 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"io"
 	"math/rand"
+	"mime/multipart"
 	"net/http"
+	"os"
+	"path/filepath"
 	"time"
 
 	"github.com/nqd/flat"
@@ -15,8 +19,105 @@ import (
 	validation "github.com/go-ozzo/ozzo-validation"
 )
 
+type MultiPartFormParams struct {
+	Key  string
+	Form any
+}
+
+type MultiPartFileParams struct {
+	Key       string
+	Filenames []string
+}
+
+// MultiPartFormRequestBuilder builds out request to include files along with optional form body
+// formKey and form parameters can be left empty/nil if body request is not needed
+func MultipartFormRequestBuilder(
+	t TestLog,
+	method,
+	url string,
+	formParams []MultiPartFormParams,
+	fileParams []MultiPartFileParams,
+	ctxVals map[any]any,
+) *http.Request {
+	t.Helper()
+
+	// Create a buffer to hold the request body
+	var buf bytes.Buffer
+
+	// Create a multipart writer to construct the multipart/form-data body
+	writer := multipart.NewWriter(&buf)
+
+	for idx, formParam := range formParams {
+		if formParam.Key == "" {
+			t.Errorf("Must set field %q at index %d for formParams", "Key", idx)
+		}
+
+		if formParam.Form != nil {
+			formBytes, err := json.Marshal(formParam.Form)
+			if err != nil {
+				t.Errorf("Error marshaling form: %s", err)
+			}
+
+			if err = writer.WriteField(formParam.Key, string(formBytes)); err != nil {
+				t.Errorf("Error writing field %q to writer: %s", formParam.Key, err)
+			}
+		}
+	}
+
+	for idx, fileParam := range fileParams {
+		if fileParam.Key == "" {
+			t.Errorf("Must set field %q at index %q for fileParams", "Key", idx)
+		}
+
+		for _, filename := range fileParam.Filenames {
+			// Open the file to be uploaded
+			file, err := os.Open(filename)
+			if err != nil {
+				t.Errorf("Error opening file: %s", err)
+			}
+			defer file.Close()
+
+			// Create the file field in the multipart form (key = "file", value = file content)
+			part, err := writer.CreateFormFile(fileParam.Key, filepath.Base(file.Name()))
+			if err != nil {
+				t.Errorf("Error creating form file: %s", err)
+			}
+
+			// Copy the contents of the file into the form data
+			_, err = io.Copy(part, file)
+			if err != nil {
+				t.Errorf("Error copying file data: %s", err)
+			}
+		}
+	}
+
+	// Close the writer to finalize the multipart form
+	writer.Close()
+
+	// Create a POST request with the correct Content-Type
+	req, err := http.NewRequest(method, url, &buf)
+	if err != nil {
+		t.Errorf("Error creating request: %s", err)
+	}
+
+	// Set the Content-Type to the value from the writer (includes boundary)
+	req.Header.Set("Content-Type", writer.FormDataContentType())
+
+	if ctxVals != nil {
+		ctx := req.Context()
+
+		for k, v := range ctxVals {
+			ctx = context.WithValue(ctx, k, v)
+		}
+
+		req = req.WithContext(ctx)
+	}
+
+	return req
+}
+
 // FormRequestBuilder takes in various params and builds a request object suited for form validation
-func FormRequestBuilder(t TestLog, method, url string, form interface{}, ctxVals map[interface{}]interface{}) *http.Request {
+func FormRequestBuilder(t TestLog, method, url string, form any, ctxVals map[any]any) *http.Request {
 	t.Helper()
 
 	formBytes, err := json.Marshal(form)
@@ -42,160 +143,6 @@ func FormRequestBuilder(t TestLog, method, url string, form interface{}, ctxVals
 	return req
 }
 
-// func ValidateFormError(t TestLog, err error, validatorMap map[string]string) {
-// 	if err != nil {
-// 		t.Helper()
-// 		var valErr validation.Errors
-
-// 		// If error is validation.Errors, determine if passed validationMap
-// 		// errors correspond with errors in validation.Errors map
-// 		//
-// 		// Else simply return err
-// 		if errors.As(err, &valErr) {
-// 			errBytes, err := err.(validation.Errors).MarshalJSON()
-
-// 			if err != nil {
-// 				t.Errorf("%+v", err)
-// 				return
-// 			}
-
-// 			var errMap map[string]interface{}
-
-// 			if err = json.Unmarshal(errBytes, &errMap); err != nil {
-// 				t.Errorf("%+v", err)
-// 				return
-// 			}
-
-// 			if validatorMap == nil {
-// 				t.Errorf("There are form errors but validatorMap is nil.  Errors: %v\n", errMap)
-// 				return
-// 			}
-
-// 			keys := make([]string, 0, len(validatorMap))
-
-// 			// Get keys from passed validatorMap to loop through
-// 			// and compare against form error
-// 			for k := range validatorMap {
-// 				keys = append(keys, k)
-// 			}
-
-// 			errObj := objx.New(errMap)
-// 			msg := ""
-
-// 			for _, key := range keys {
-// 				// Use given key to get value from our errObj
-// 				errObjVal := errObj.Get(key)
-
-// 				// If value is nil in errObj, this tells us that given key
-// 				// is not in errObj so add to our error msg
-// 				//
-// 				// Else compare values
-// 				if errObjVal.IsNil() {
-// 					msg += fmt.Sprintf("Key \"%s\" was not found in form errors\n", key)
-// 				} else {
-// 					if errObjVal.Str() != validatorMap[key] {
-// 						msg += fmt.Sprintf(
-// 							"Key \"%s\" threw err: \"%s\" \n expected: \"%s\"\n",
-// 							key,
-// 							errObjVal.Str(),
-// 							validatorMap[key],
-// 						)
-// 					}
-
-// 					// walkMapFunc helps us remove key from errMap so at
-// 					// the end of the key loop, if there is any values left,
-// 					// those will be given back in an error as those will be
-// 					// errors that weren't in the validatorMap passed
-// 					var walkMapFunc func(string)
-
-// 					walkMapFunc = func(mapKey string) {
-// 						mapLvls := strings.Split(mapKey, ".")
-
-// 						// getMapKey is anon util function that helps us get the key
-// 						// for the map above current mapKey
-// 						getMapKey := func() string {
-// 							k := ""
-
-// 							for i, v := range mapLvls[:len(mapLvls)-1] {
-// 								k += v
-
-// 								if i != len(mapLvls)-2 {
-// 									k += "."
-// 								}
-// 							}
-
-// 							return k
-// 						}
-
-// 						// If errObj with mapKey currently has str value, we know we want to
-// 						// delete this entry
-// 						//
-// 						// Else mapKey should be for entry for a map[string]interface{}
-// 						if errObj.Get(mapKey).IsStr() {
-// 							// If mapLvls is 1, then current key is entry in our errMap so simply delete
-// 							//
-// 							// Else current key is for nested map so we must traverse
-// 							if len(mapLvls) == 1 {
-// 								delete(errMap, mapKey)
-// 							} else {
-// 								// Get key for map above current key entry and remove entry from that map
-// 								k := getMapKey()
-// 								m := errObj.Get(k).Data().(map[string]interface{})
-// 								mapProperty := mapLvls[len(mapLvls)-1]
-// 								delete(m, mapProperty)
-
-// 								// Pass in mapKey that got us map "m" into walkMapFunc for recursion
-// 								walkMapFunc(k)
-// 							}
-// 						} else {
-// 							// Get map[string]interface{} from mapKey
-// 							tmpMap := errObj.Get(mapKey).Data().(map[string]interface{})
-
-// 							// If map has no entries, begin trying to remove
-// 							if len(tmpMap) == 0 {
-// 								// If mapLvls is 1, then current map is entry in our errMap
-// 								// so simple remove
-// 								//
-// 								// Else get map above current map and delete current map from
-// 								// above map entry
-// 								if len(mapLvls) == 1 {
-// 									delete(errMap, mapKey)
-// 								} else {
-// 									k := getMapKey()
-// 									m := errObj.Get(k).Data().(map[string]interface{})
-// 									mapProperty := mapLvls[len(mapLvls)-1]
-// 									delete(m, mapProperty)
-
-// 									// Pass in mapKey that got us map "m" into walkMapFunc for recursion
-// 									walkMapFunc(k)
-// 								}
-// 							}
-// 						}
-// 					}
-
-// 					walkMapFunc(key)
-// 				}
-// 			}
-
-// 			if len(errMap) > 0 {
-// 				msg += fmt.Sprintf("\n\n These are form errors not found in validatorMap: %v\n", errMap)
-// 			}
-
-// 			if msg != "" {
-// 				t.Errorf(msg)
-// 				return
-// 			}
-// 		} else {
-// 			t.Errorf("%+v", err)
-// 			return
-// 		}
-// 	} else {
-// 		if validatorMap != nil {
-// 			t.Errorf("There are no form errors but validatorMap was passed")
-// 		}
-// 	}
-// }
-
 // ValidateFormError determines whether the error passed is a form error
 // and whether it returns the expected errors with the validatorMap parameter
 //
@@ -212,13 +159,12 @@ func ValidateFormError(t TestLog, err error, validatorMap map[string]string) {
 		// Else simply return err
 		if errors.As(err, &valErr) {
 			errBytes, err := err.(validation.Errors).MarshalJSON()
-
 			if err != nil {
 				t.Errorf("%+v", err)
 				return
 			}
 
-			var errMap map[string]interface{}
+			var errMap map[string]any
 
 			if err = json.Unmarshal(errBytes, &errMap); err != nil {
 				t.Errorf("%+v", err)
